+++ cmsaf/NEWS.md
@@ -1,8 +1,12 @@
 # Release notes for cmsaf
 
+## 3.6.1
+- Improve color choice for significance mask
+
 ## 3.6.0
 - Added option to change font size in plots (works with cmsafvis >= 1.2.10)
 - Added message for file chooser, which might open behind Toolbox window
+- Add visualize options to plot significance of trend
 
 ## 3.5.2
 - Minor fix of test